perf(core): optimize BHVec10240 bundling via bit-sliced AVX2/NEON
